+++ components/org.wso2.carbon.uis/src/main/java/org/wso2/carbon/uis/api/App.java
@@ -19,10 +19,10 @@
 package org.wso2.carbon.uis.api;
 
 import org.wso2.carbon.uis.api.exception.PageNotFoundException;
+import org.wso2.carbon.uis.api.exception.PageRedirectException;
 import org.wso2.carbon.uis.api.http.HttpRequest;
 import org.wso2.carbon.uis.api.util.Multilocational;
 import org.wso2.carbon.uis.api.util.Overridable;
-import org.wso2.carbon.uis.internal.exception.PageRedirectException;
 import org.wso2.carbon.uis.internal.impl.OverriddenApp;
 
 import java.util.Collection;
