+++ components/org.wso2.carbon.uis/src/main/java/org/wso2/carbon/uis/internal/RequestDispatcher.java
@@ -22,10 +22,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.carbon.uis.api.App;
 import org.wso2.carbon.uis.api.exception.HttpErrorException;
+import org.wso2.carbon.uis.api.exception.PageRedirectException;
 import org.wso2.carbon.uis.api.exception.UISRuntimeException;
 import org.wso2.carbon.uis.api.http.HttpRequest;
 import org.wso2.carbon.uis.api.http.HttpResponse;
-import org.wso2.carbon.uis.internal.exception.PageRedirectException;
 import org.wso2.carbon.uis.internal.http.ResponseBuilder;
 import org.wso2.carbon.uis.internal.io.StaticResolver;
