+++ components/org.wso2.carbon.uis/src/test/java/org/wso2/carbon/uis/api/AppTest.java
@@ -21,8 +21,8 @@ package org.wso2.carbon.uis.api;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import org.wso2.carbon.uis.api.exception.PageNotFoundException;
+import org.wso2.carbon.uis.api.exception.PageRedirectException;
 import org.wso2.carbon.uis.api.http.HttpRequest;
-import org.wso2.carbon.uis.internal.exception.PageRedirectException;
 import org.wso2.carbon.uis.internal.impl.HtmlPage;
 
 import java.util.ArrayList;
